feat: optimized for speed using cpi jit decorator

+++ try2/operations.py
@@ -5,5 +5,5 @@ from numba import njit
 def unit_vector(vector):
     norm = np.linalg.norm(vector)
     if norm == 0:
-        return vector
-    return vector / norm
+        return vector.astype(np.float32)
+    return (vector / norm).astype(np.float32)
